# book-social-network/src/main/java/com/xuntrng/book_social_network/auth/AuthenticationResponse.java
package com.xuntrng.book_social_network.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}